extract all insights files into their own folder
