+++ packages/cli/src/databases/repositories/insights-metadata.repository.ts
@@ -1,7 +1,7 @@
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
 
-import { InsightsMetadata } from '../entities/insights-metadata';
+import { InsightsMetadata } from '../../insights/entities/insights-metadata';
 
 @Service()
 export class InsightsMetadataRepository extends Repository<InsightsMetadata> {
