+++ packages/cli/src/insights/entities/__tests__/insights-by-period.test.ts
