+++ packages/cli/src/insights/entities/insights-by-period.ts
@@ -1,7 +1,6 @@
 import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from '@n8n/typeorm';
 import { UnexpectedError } from 'n8n-workflow';
 
-import { datetimeColumnType } from './abstract-entity';
 import type { PeriodUnits } from './insights-shared';
 import {
 	isValidPeriodNumber,
@@ -11,6 +10,7 @@ import {
 	PeriodUnitToNumber,
 	TypeToNumber,
 } from './insights-shared';
+import { datetimeColumnType } from '../../databases/entities/abstract-entity';
 
 @Entity()
 export class InsightsByPeriod extends BaseEntity {
