+++ packages/cli/src/insights/entities/insights-raw.ts
@@ -4,8 +4,11 @@ import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from '@n8n/typeorm
 import { DateTime } from 'luxon';
 import { UnexpectedError } from 'n8n-workflow';
 
-import { datetimeColumnDefault, datetimeColumnType } from './abstract-entity';
 import { isValidTypeNumber, NumberToType, TypeToNumber } from './insights-shared';
+import {
+	datetimeColumnDefault,
+	datetimeColumnType,
+} from '../../databases/entities/abstract-entity';
 
 export const { type: dbType } = Container.get(GlobalConfig).database;
