+++ packages/cli/src/insights/entities/insights-shared.ts
