+++ packages/cli/src/insights/insights.config.ts
@@ -1,4 +1,4 @@
-import { Config, Env } from '../decorators';
+import { Config, Env } from '@n8n/config/src/decorators';
 
 @Config
 export class InsightsConfig {
