+++ packages/cli/test/integration/shared/db/insights.ts
@@ -1,13 +1,13 @@
 import { Container } from '@n8n/di';
 import type { DateTime } from 'luxon';
 
-import { InsightsByPeriod } from '@/databases/entities/insights-by-period';
-import { InsightsMetadata } from '@/databases/entities/insights-metadata';
-import { InsightsRaw } from '@/databases/entities/insights-raw';
 import type { WorkflowEntity } from '@/databases/entities/workflow-entity';
 import { InsightsByPeriodRepository } from '@/databases/repositories/insights-by-period.repository';
 import { InsightsMetadataRepository } from '@/databases/repositories/insights-metadata.repository';
 import { InsightsRawRepository } from '@/databases/repositories/insights-raw.repository';
+import { InsightsByPeriod } from '@/insights/entities/insights-by-period';
+import { InsightsMetadata } from '@/insights/entities/insights-metadata';
+import { InsightsRaw } from '@/insights/entities/insights-raw';
 
 import { getWorkflowSharing } from './workflows';
